requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.3.0
duckduckgo-search>=6.2.0
